audioflinger: pause direct output when buffer timeout
Direct output can be paused when track underrun is observed
for the first time, so this can interrupt the playback even
when there is PCM data cached in the driver. Also the track
may be ready before the next prepareTracks, the playback
should be continued.
To avoid the unnecessary pause, only do HAL pausing when
track is going to be removed due to buffer timeout.

Change-Id: I3707d883d0daa4ffd2baa9bd69aebf4bf9ddef4e
CRs-Fixed: 1094472

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>