CameraService: Consider pids before priority when evicting a client.
We are occasionally seeing differences in priority scores for the same
pid when a process tries to open a second camera (or the same one twice
without closing it). While this is a problem by itself, checking the pid
before the priority scores would ensure the correct behavior for
consecutive calls to opening the camera (closing the first camera
connection and opening up a second one). This is especially true for
cases where oomScoreOffset is nonzero, since it will only be considered
for the new client.

Bug: 233743526
Test: Ran custom cameraservice_test before and after fix.
Change-Id: Ie6b678681d1141e077670e627f9c4244fbdf1287
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>